requirements.txt
build==1.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
docutils==0.20.1
filelock==3.13.3
fsspec==2024.3.1
idna==3.6
importlib_metadata==7.1.0
jaraco.classes==3.3.1
jaraco.context==4.3.0
jaraco.functools==4.0.0
jeepney==0.8.0
Jinja2==3.1.3
keyring==25.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
nh3==0.2.15
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pillow==10.2.0
pkginfo==1.10.0
pycparser==2.21
Pygments==2.17.2
pyproject_hooks==1.0.0
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.3
sympy==1.12
tomli==2.0.1
torch==2.2.1
triton==2.2.0
twine==5.0.0
typing_extensions==4.10.0
urllib3==2.2.1
zipp==3.18.1
